spec: fix endian opcodes
From looking at the LLVM backend I see that all byteswap opcodes use the ALU
class, and they use the source bit to decide whether to do the swap. If the
source bit is set (as in be64) then the VM will byteswap iff it is running on a
little endian machine. If the source bit is unset, then it will swap iff it is
running on a big endian machine. However, it looks like LLVM actually decides
the endianness of the target machine at compile time.