vm: implement CALL instruction
Before loading code into the VM, the application should use the ubpf_register
API to associate C functions to integer indices. The eBPF program can then use
these indicies as the immediate in the CALL to pick which function to call.

The 'name' argument is currently unused, but will be used later by the ELF
loader.