Add BPF_J{LT,LE,SLT,SLE} instructions
See kernel patches merged in commit 078295fb9af5. Those new instructions
were added for performance reason to the kernel. Add them to uBPF.

This commit adds the new instructions for the interpreter, the JIT, the
assembler and the disassembler. Unit tests are also added for these
instructions.

As rbpf took its unit tests from uBPF, the unit tests in this commit are
also consistent with the associated tests for the new instructions for
rbpf.