Remove obsolete unary_function
Since std::variant is used, C++17 is required, since std::variant was added
in C++17 (see https://en.cppreference.com/w/cpp/utility/variant)

However std::unary_function was removed in C++17, (see
https://en.cppreference.com/w/cpp/utility/functional/unary_function and
it was originally deprecated in C++11 with rationale in
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2010/n3145.html)
and so cannot be safely used in code that needs to compile with C++17.
This causes the CRAB library to fail to compile with some compilers,
and any compiler it works with is arguably non-conformant with the C++
standard.

Per https://stackoverflow.com/questions/22386882/why-have-unary-function-binary-function-been-removed-from-c11
the correct resolution is to simply delete the inheritance.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>