Fix heap corruption
The indices[] array was incorrectly resized downwards when it should
only be resized to be bigger.  Compare line 93 at
https://github.com/seahorn/crab/blob/master/include/crab/domains/graphs/util/Heap.h
which uses a custom growTo() function instead of resize(), and line
172 of
https://github.com/seahorn/crab/blob/master/include/crab/domains/graphs/util/Vec.h
avoids resizing downwards, which the ebpf_verifier version didn't do.

The symptom is that depending on how vector.resize() is implemented
on the platform, the code would either
a) index into memory past the end of the array in percolateUp,
   which should be illegal but isn't caught
   (this is presumably what happens in the current CI), OR
b) crash due to accessing invalid memory

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>