Replace map_def with EbpfMapDescriptor
This is the next step in reducing use of code in the gpl subdirectory.

The verifier today never uses MapType for anything, so leave it
raw like in bpf_load_map_def in asm_files.cpp, to allow swapping
the order of map fd replacement vs. verification at some point
in the future.

program_info and raw_program will move in a later PR since they require
BpfProgType to be handled first.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>