Move maps section parsing to platform-specific code
Since the size of a record in the maps section can vary by platform
and even by version of the same platform.  For example, the generic-ebpf
project only has 5, not 7, fields in the struct, and even Linux has
extended the struct over time it appears.

Also print error on invalid maps section instead of silently crashing
due to an exception.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>