Look up map descriptor by fd rather than using bits
This lets the code work with either a mock fd like create_map_crab uses
or an actual fd, without needing to know the difference since the
map_descriptors are stored in the program_info.

create_map_linux() is moved from src/main/linux_verifier.cpp to
src/linux/linux_platform.cpp to allow for the possibility of using
the prevail verifier on an ebpf program that has real fd's in it,
i.e., creating real maps but then just calling the prevail verifier
instead of the linux verifier.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>