Replace Weak Topological Ordering code with MIT licensed version
Replace the WTO implementation with a new one written entirely based off
the specification in the paper.

Previously the code kept track of how many times the cycle is visited,
but the paper does not mention it, and so the new implementation does not
keep track of it.  (Also it turns out that information was not used
anyway.)

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>