Make wronghelper.o fail unmarshal()
The basic idea is that it is up to the platform to determine whether
a given helper ID is valid for a given program type.  This PR updates
the Linux platform implementation to verify that the helper's ctx (if
any) has the same context_descriptor as the program type.

The platform might still have further constraints that were not implemented
before and this PR does not add.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>