Move external structures to a C compatible header
Some definitions must be consistent with, and ideally directly used by,
an eBPF runtime that exposes hooks and helpers.  C++ code is typically
not used in kernels, and so such enums/structs should be in a header
that can be used directly by an eBPF runtime written in C, such as
the ebpf-for-windows project which uses the PREVAIL verifier as a
submodule.

This PR enables the runtime project to use the header directly rather
than duplicating definitions in the runtime project which would
potentially lead to inconsistencies.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>