Add dependabot support
The security principle is that one should try to
upgrade as soon as fixes are available in dependencies.

With this config, the github dependabot will automatically open
a pull request to update a dependency when one is available.
apt dependencies are not yet supported by dependabot, just pip,
git submodules, github actions, etc. as listed at
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#package-ecosystem

Signed-off-by: Dave Thaler <dthaler@microsoft.com>