Deprecate legacy packet access instructions
The BPF ISA specification does not define them, and just says
"these instructions are deprecated and should no longer be used".
To comply with that, this PR disables support for legacy packet access
instructions by default.

If some other project depends on them, options.legacy can be
set to true.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>