Implement callx instruction
These instructions are generated by both clang (under -O0 or -O1)
and gcc (if the experimental -mxbpf option is passed to the compiler),
but are not supported by Linux.

Per mailing list discussion at
https://mailarchive.ietf.org/arch/msg/bpf/CDQjTO8R8gdPdfeKVnoxWco8_Lw/
the intent is to support them eventually, and with this PR,
ebpf-for-windows can support them now.  This will also unblock
the ability to use versions of clang later than clang-11 by using -O1
instead of -O2 (which generates correlated branches PREVAIL can't
deal with).

The mailing list discussion at
https://mailarchive.ietf.org/arch/msg/bpf/Vx1H3ViPUWoGKNssCO22lOIjyXU/
agreed that inst.dst is where to put the register to use going forward
and clang v.19 will do so but earlier versions of clang used inst.imm.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>