Put bpf2bpf caller in a named section
Libbpf logic is that if there are multiple programs then any programs in the .text section
are subprograms to be skipped when enumerating programs.

So the sample was updated to follow the convention that the non-subprogram is in a section
other than ".text", allowing testing the ability to load programs that call subprograms in
another section.  This PR updates the test accordingly and fixes a bug
hit that prevented bpf2bpf calls across sections.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>