Fix bpf2bpf bug when multiple branches in a subprogram go to exit
The bug was that the successor label could be processed multiple
times if there were multiple predecessors, resulting in multiple
exit instructions appearing in the block, which in turn would
result in restore_callee_saved_registers() being called multiple
times, giving incorrect results.

Added a YAML test case for this.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>