Skip relocations to sections when parsing elf
When programs contain a section that contain data, e.g., read-only data
stored in .rodata section, relocations to the data show up as
relocation to the corresponding section, which results in "Unresolved
external symbol" error. We can skip when a relocation to a section is
encountered to avoid such errors.

Signed-off-by: Ameer Hamza <ah18r@fsu.edu>